Move PageHero to hurumap/core
