+++ packages/hurumap-core/src/PageHero/PageHero.snap.js
