+++ packages/hurumap-core/src/RichHeader/RichHeader.js
