+++ packages/hurumap-core/src/RichHeader/index.js
