+++ packages/hurumap-core/src/index.js
@@ -16,3 +16,5 @@ export { default as HowItWorks } from "./HowItWorks";
 export { default as LineClampedRichTypography } from "./LineClampedRichTypography";
 export { default as Loading } from "./Loading";
 export { default as Menu } from "./Menu";
+export { default as RichHeader } from "./RichHeader";
+export { default as PageHero } from "./PageHero";
